Change expected behavior around unknown enum values
Unknown enum values should never lead to errors at the deserialization
layer, and should instead be persisted to the application layer.

This change to the guidelines for enums comes after some consideration
of the way that the official EO client handles enums (as raw numbers).
Following the old guidelines, a game server could easily break any
client based on a library generated from eo-protocol. It would be as
simple as sending an out-of-bounds value for a field in all of the pub
files and/or maps, causing the deserializer to error out.