Add `blob` type
This type is conceptually different from array of `byte`, though it
should be serialized identically to an unbounded byte array.

The change is primarily so code generation can distinguish between
"array of bytes" and "opaque blob".
For example, an `eolib` implementation can consistently treat all
numeric arrays as lists of the standard integer type, while treating
`blob` as an array of bytes.